# backend/internal/apiserver/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/tenant:
        get:
            tags:
                - TenantService
            description: |-
                ListTenants 列出租户（平台上下文，跨租户全量）。
                   REST: GET /v1/tenant
            operationId: TenantService_ListTenants
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListTenantsResponse'
        post:
            tags:
                - TenantService
            description: |-
                CreateTenant 创建租户。
                   REST: POST /v1/tenant
            operationId: TenantService_CreateTenant
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTenantRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateTenantResponse'
    /v1/tenant/{id}:
        get:
            tags:
                - TenantService
            description: |-
                GetTenant 取租户。
                   REST: GET /v1/tenant/{id}
                 注意 REST 用单数段：DefaultHTTPObject 取首资源段（"tenant"），与 gRPC
                 DefaultGRPCObject("TenantService/...")="tenant" 同源（P9 归一化，策略单写）。
            operationId: TenantService_GetTenant
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetTenantResponse'
        put:
            tags:
                - TenantService
            description: |-
                UpdateTenant 更新租户。
                   REST: PUT /v1/tenant/{id}
            operationId: TenantService_UpdateTenant
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTenantRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateTenantResponse'
        delete:
            tags:
                - TenantService
            description: |-
                DeleteTenant 删除租户。
                   REST: DELETE /v1/tenant/{id}
            operationId: TenantService_DeleteTenant
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteTenantResponse'
...
components:
    schemas:
...
        CreateTenantRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                remark:
                    type: string
        CreateTenantResponse:
            type: object
            properties:
                tenant:
                    $ref: '#/components/schemas/Tenant'
...
        DeleteTenantResponse:
            type: object
            properties:
                deleted:
                    type: string
...
        GetTenantResponse:
            type: object
            properties:
                tenant:
                    $ref: '#/components/schemas/Tenant'
...
        ListTenantsResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tenant'
                total:
                    type: integer
                    description: 注意用 uint32：proto3 JSON 把 int64/uint64 编码为字符串，会破坏常规 JSON 客户端。
                    format: uint32
...
        Tenant:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - STATUS_ON
                        - STATUS_OFF
                        - STATUS_FREEZE
                    type: string
                    format: enum
                remark:
                    type: string
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
            description: |-
                Tenant 租户。ID 即租户编码（business code，如 "platform"、"t-acme"），
                 与 bald P8 的 tenant_id 同一命名空间——业务租户创建后即作为 P8 隔离维度使用。
...
        UpdateTenantRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - STATUS_ON
                        - STATUS_OFF
                        - STATUS_FREEZE
                    type: string
                    format: enum
                remark:
                    type: string
        UpdateTenantResponse:
            type: object
            properties:
                tenant:
                    $ref: '#/components/schemas/Tenant'
...